safesocket: add ability for Darwin clients to set explicit credentials
updates tailscale/corp#25687

The darwin gui clients now support XPC and the keychain
for passing user credentials securely between the gui process
and tailscaled.  Clients that can communicate directly with
tailscaled,  via XPC or the keychain, are now expected to call
SetCredentials and supply credentials explicitly, fixing issues
with the cli breaking if the current user cannot read the
contents of /Library/Tailscale due to group membership
restrictions.  This matches how those clients source and supply
credentials to the localAPI http client.

Non-platform-specific code that has traditionally been in
the client is moved to safesocket.

/Libraray/Tailscaled/sameuserproof is now restricted to
users in the admin group.

Signed-off-by: Jonathan Nobels <jonathan@tailscale.com>