tailcfg: add a new TailscaleFunnelUsageDetected method
In capver 113 we added new Hostinfo.IngressEnabled field to send to control
info if a node has any active funnel endpoint.
This is on top of the existing Hostinfo.WireIngress field that sends
whether the user intends to use funnel (but it's not neccessarily active).
We are now in a state where control will be parsing Hostinfos
that can use the new (IngressEnabled) field to more correctly
determine if funnel is really enabled as well as Hostinfos
that must use the the old Hostinfo.WireIngress field.

This PR adds a new TailscaleFunnelUsageDetected method
that can be used by control to determine if funnel is enabled
for Hostinfos for clients of capver <= 112.
Once this merges and changes are made control-side to use this method
for <= capver 112 Hostinfos, we can change the
currently used TailscaleFunnelEnabled field to use the new Hostinfo.IngressEnabled field.

Updates tailscale/tailscale#11572
Updates tailscale/corp#25931

Signed-off-by: Irbe Krumina <irbe@tailscale.com>