feature/capture: Fix dissector zero-division bug
The first byte in DISCO metadata is used for flags. It has only one flag
and is set to `0x01` if the source address is the DERP magic adress. See
[`disco/pcap.go`](https://github.com/tailscale/tailscale/blob/986daca5eeeffa04bdb184d1ee13f70d04d33ff1/disco/pcap.go#L25-L28).

The dissector would fail to dissect DISCO frames with the error message
`Lua Error: [path/to]/.local/lib/wireshark/plugins/ts-dissector.lua:2: attempt to perform 'n%0'`.

Replaced the buggy `hasbit` function with a new `is_from_derp` function
to keep the implementation simple and readable.

Tested on macOS with Wireshark 4.4.0.

Fixes #15182

Signed-off-by: Ole Kristian Pedersen <okpedersen@gmail.com>