Improve description of limiting logged messages
Previosuly, logging might itself cause a logged message in the form
    > tailscaled: [RATELIMIT] format("%s: %s %d %s\n%s")

That looks like a mis-formatted message, and it has no verb or
obvious cause. It looks like a dire problem of something happening
too often, when really it is a space-saving device for log files.
Logging needed rate limiting, not a rate limit needed logging.

This changes the function-call-looking message to tell what is
actually happening: We are suppressing further log messages.